requirements.txt
kafka
kafka-shell
PyPubSub
mysql-connector-python
requests
Flask==0.10.1
db-sqlite3
Flask-RESTful